requirements.txt
PyYAML~=5.4.1
xmltodict~=0.12.0
